update libusb and libuvc to fix some issues and to improve compatibility with some cameras add some functions including request error code control functions described at the section 4.2.1.2 of UVC 1.5 class specification.

+++ library/jni/libusb/libusb/os/android_usbfs.h
@@ -42,7 +42,7 @@ struct usbfs_ctrltransfer {
 struct usbfs_bulktransfer {
 	/* keep in sync with usbdevice_fs.h:usbdevfs_bulktransfer */
 	unsigned int ep;
-	size_t len;				// XXX changed from unsigned int len;
+	unsigned int len;
 	unsigned int timeout;	/* in milliseconds */
 
 	/* pointer to data */
@@ -76,14 +76,14 @@ enum usbfs_urb_type {
 };
 
 struct usbfs_iso_packet_desc {
-	size_t length;			// XXX changed from unsigned int length;
-	size_t actual_length;	// XXX changed from unsigned int actual_length;
+	unsigned int length;
+	unsigned int actual_length;
 	unsigned int status;
 };
 
-#define MAX_ISO_BUFFER_LENGTH		32768
-#define MAX_BULK_BUFFER_LENGTH		16384
-#define MAX_CTRL_BUFFER_LENGTH		4096
+#define MAX_ISO_BUFFER_LENGTH		32768	// 32 x 1024 = 32KB
+#define MAX_BULK_BUFFER_LENGTH		16384	// 16 x 1024 = 16KB
+#define MAX_CTRL_BUFFER_LENGTH		4096	// 4 x 1024 = 4KB
 
 struct usbfs_urb {
 	unsigned char type;
@@ -91,8 +91,8 @@ struct usbfs_urb {
 	int status;
 	unsigned int flags;
 	void *buffer;
-	size_t buffer_length;	// XXX changed from int buffer_length;
-	size_t actual_length;	// XXX changed from int actual_length;
+	int buffer_length;
+	int actual_length;
 	int start_frame;
 	union {
 		int number_of_packets;	/* Only used for isoc urbs */
